[server] change path name helper
