+++ myBitboxMining/webnode-mining/modelMgrs/OrderManager.js
@@ -20,9 +20,6 @@ const OrderTable = sequelize.define('orders', {
     updateAt : Sequelize.DATE
 });
 
-// class helper
-const FileHelper = require('../private/js/FileHelper');
-
 module.exports = {
 
     /**
