+++ myBitboxMining/webnode-mining/modelMgrs/UserManager.js
@@ -25,7 +25,7 @@ const UserTable = sequelize.define('user', {
 UserTable.removeAttribute('id');
 
 // class helper
-const FileHelper = require('../private/js/FileHelper');
+const FileHelper = require('../global/FileHelper');
 
 // require module
 const moment = require('moment');
